Chore: Move root test scripts to scripts directory
Relocate test scripts from project root to scripts/ for better organization:
- test_api_crud.py
- test_api_crud.sh
- test_api_crud_v2.py
- test_comprehensive_crud.py

+++ apps/backend/app/domains/shopping/schemas/cart_schemas.py
@@ -26,7 +26,7 @@ class CartProductResponse(BaseModel):
     id: uuid.UUID
     sku: Optional[str] = None
     name: str
-    price: Optional[int] = None
+    price: Optional[float] = None
     images: List[ProductImageResponse] = []
     stock_quantity: int
 
@@ -40,7 +40,7 @@ class CartItemResponse(BaseModel):
     cart_id: uuid.UUID
     product_id: uuid.UUID
     quantity: int
-    unit_price: Optional[int] = None
+    unit_price: Optional[float] = None
     notes: Optional[str] = None
     substitution_allowed: bool
     product: CartProductResponse
